requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
inflection==0.5.1
iso8601==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
m3u8==3.6.0
Markdown==3.4.4
outcome==1.2.0
Pillow==10.0.1
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.4
uritemplate==4.1.1
urllib3==2.0.5
wsproto==1.2.0
